Rename classes as part of API streamline
See commit 944cf7a for the proper diff
